Add UI Commit
Add normal HealthBar
Add Skybox
Create KillsCount (DONT FORGET TO GETINT FOR KILLS!!!!!!!!!!!)
Add Image for Main Menu and Loading

+++ Assets/ScriptableObjects/PlayerData.cs
@@ -5,5 +5,5 @@ public class PlayerData : ScriptableObject
 {
     public int health;
     public int damage;
-    public int moveSpeed;
+    public float moveSpeed;
 }
